[FIX] Previous choices were bad
