+++ Week5/frontend/packages/nextjs/app/lottery/page.tsx
@@ -3,31 +3,12 @@
 import type { NextPage } from "next";
 import { LoadContractAddress } from "./components/load-contract";
 import { BuyTokens } from "./components/buy-tokens";
-import { useEffect, useState } from "react";
+import { useState } from "react";
 import { DelegateAllowance } from "./components/delegate-allowance";
 import { RedeemTokens } from "./components/redeem-tokens";
-import { usePublicClient } from "wagmi";
-import { abi as lotteryAbi } from "../../../../../lottery-contract/artifacts/contracts/Lottery.sol/Lottery.json"
-
-
 const Lottery: NextPage = () => {
     const [address, setAdress] = useState("");
     const [tokenAddress, setTokenAddress] = useState<string | null>(null);
-    const client = usePublicClient();
-
-    useEffect(
-        () => {
-            client?.readContract({
-                abi: lotteryAbi,
-                address: address,
-                functionName: "paymentToken",
-            }).then((addr) => setTokenAddress(addr as string)).catch((e: Error) => {
-                console.log("ERROR occured : ", e.message)
-                throw e;
-            })
-
-        }, [address]
-    )
 
     const debugMode = process.env.NEXT_PUBLIC_CHAIN_ENV == "hardhat"
     let blockExplorer = "";
@@ -45,8 +26,8 @@ const Lottery: NextPage = () => {
                         <br /> Check{" "}
                     </p>
 
-                    <LoadContractAddress setAddress={setAdress} address={address} />
-                    {address && (<>
+                    <LoadContractAddress setAddress={setAdress} setTokenAddress={setTokenAddress} address={address} />
+                    {address && tokenAddress && (<>
                         <BuyTokens address={address} blockExplorer={blockExplorer} />
                         <DelegateAllowance address={address} tokenAddress={tokenAddress} blockExplorer={blockExplorer} />
                         <RedeemTokens address={address} blockExplorer={blockExplorer} />
